Hw5

+++ SpringProject/link-parser/src/main/java/ru/tinkoff/edu/java/parser/results/StackOverflowParseResult.java
@@ -7,7 +7,7 @@ import lombok.ToString;
 @ToString
 public class StackOverflowParseResult extends ParseResult {
     @Getter
-    private final int questionId;
+    private final long questionId;
     public StackOverflowParseResult(@NonNull String url, int questionId) {
         super(url);
         this.questionId = questionId;
